% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketStatsAPI.R
\name{MarketStatsAPI}
\alias{MarketStatsAPI}
\title{Get day-by-day general stats}
\usage{
MarketStatsAPI(date1 = "10daysAgo", date2 = "today",
  client_id = NULL, shop_id = NULL, token = NULL)
}
\arguments{
\item{date1}{Start date}

\item{date2}{End date}

\item{client_id}{Your API token, see \code{\link{https://tech.yandex.ru/market/partner/doc/dg/concepts/authorization-docpage/}}}

\item{shop_id}{Your shop ID, see \code{\link{https://partner.market.yandex.ru/?list=yes}}}

\item{token}{Your API token, see \code{\link{https://tech.yandex.ru/market/partner/doc/dg/concepts/authorization-docpage/}}}
}
\description{
Function to get day-by-day general stats
}
\examples{
MarketStatsAPI()
}
