% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarketStatsOffersAPI30.R
\name{MarketStatsOffersAPI30}
\alias{MarketStatsOffersAPI30}
\title{Get stats by offers}
\usage{
MarketStatsOffersAPI30(date1 = "10daysAgo", date2 = "today",
  client_id = NULL, shop_id = NULL, token = NULL, feedId = NULL)
}
\arguments{
\item{date1}{Start date}

\item{date2}{End date}

\item{client_id}{Your API token, see \code{\link{https://tech.yandex.ru/market/partner/doc/dg/concepts/authorization-docpage/}}}

\item{shop_id}{Your shop ID, see \code{\link{https://partner.market.yandex.ru/?list=yes}}}

\item{token}{Your API token, see \code{\link{https://tech.yandex.ru/market/partner/doc/dg/concepts/authorization-docpage/}}}

\item{feedId}{You can check it in Pricelabs}
}
\description{
Function to get stats by offers. Maximum period is 30 days!
}
\examples{
MarketStatsOffersAPI30()
}
